chore: move error components to features dir
